BS stepper: correct exponents for optimal h
Becoming suspicious by the difference of the exponents used for computing the
new step size in the BS and BS denseout stepper (see
0f943fbf8b303a8083876bb1fba677d4c4677417) I checked again the Hairer book and
I'm now convinced there was a mistake in our implementation and both
steppers should use 1/(2*k+1) as exponent. The background is the this exponent
represents the order of the error of the k-th iteration, and this order is
always 2k+1, independent of the interval_sequence. This error is computed from
the difference of the k-th and k-1 - th iteration, which have the orders 2k+2
2k respectively, which means the computed error has order 2k+1.