some approach to use arrays of eigen vectors
Using arrays (or vectors) of Eigen matrices is currently not working with
controlled stepper due to contradicting usages of std::abs. The right approach
for this situation would probably be to introduce some "composed" algebra like
compose<array_algebra, vector_space_algebra> such that for the outer array the
array_algebra is used, while the inner Eigen matrices are dealt with by the
vector_space_algebra. The fix introduced here only applies to this very
specific situation without introducing a composed algebra.